refactor: change navigation route to use universal nft
